fix(ide): deduplicate PRESENCE_DOT constant into keeper-presence-store
